using byte for identifiers

+++ source/com/BombingGames/Caveland/Game/ChunkGenerator.java
@@ -9,7 +9,7 @@ import com.BombingGames.WurfelEngine.Core.Map.Generator;
 public class ChunkGenerator implements Generator {
 
 	@Override
-	public int generate(int x, int y, int z) {
+	public byte generate(int x, int y, int z) {
 		if (z<3) return 2;
 		if (z==3) return 1;
 		return 0;
