+++ source/com/BombingGames/Caveland/GameObjects/TFlint.java
@@ -56,7 +56,7 @@ public class TFlint extends Collectible {
 	public void action() {
 		super.action();
 		Controller.getSoundEngine().play("hiss", getPosition());
-		setValue(5);
+		setValue((byte) 5);
 		lit = true;
 		timer = TIMETILLEXPLOSION;
 	}
