+++ source/com/BombingGames/WurfelEngine/Core/Gameobjects/Sea.java
@@ -39,7 +39,7 @@ public class Sea extends AnimatedBlock {
      *
      * @param id
      */
-    public Sea(final int id) {
+    public Sea(final byte id) {
         super(id,new int[]{200,200,200,200}, true, true);
 		setBounce(true);
     }
