+++ source/com/BombingGames/WurfelEngine/Core/Map/AbstractMap.java
@@ -51,7 +51,7 @@ public abstract class AbstractMap implements Cloneable {
 	private boolean modified = true;
 	private ArrayList<LinkedWithMap> linkedObjects = new ArrayList<>(3);//camera + minimap + light engine=3 minimum
 	private float gameSpeed;
-	private final CoreData groundBlock = new CoreData(CVar.get("groundBlockID").getValuei()); //the representative of the bottom layer (ground) block
+	private final CoreData groundBlock = new CoreData((byte) CVar.get("groundBlockID").getValuei()); //the representative of the bottom layer (ground) block
 	/**
 	 * holds the metadata of the map
 	 */
